pyproject.toml
[project]
name = "ai-agents-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "firecrawl-py>=2.16.3",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "mcp[cli]>=1.11.0",
    "openai-agents[voice]>=0.2.0",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "pyttsx3>=2.99",
    "streamlit>=1.46.1",
    "twilio>=9.6.5",
]
